ai-platform-validator/requirements.txt
requests>=2.31.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
cryptography>=41.0.0
aiohttp>=3.9.0

